Just getting it to compile

+++ Frameworks/XADMaster.framework/Versions/A/Headers/XADArchive.h
@@ -1,8 +1,8 @@
 #import <Foundation/Foundation.h>
 
 #import "XADArchiveParser.h"
-#import "XADUnarchiver.h"
 #import "XADException.h"
+#import "XADUnarchiver.h"
 
 typedef int XADAction;
 //typedef off_t xadSize; // deprecated
@@ -20,136 +20,135 @@ extern NSString *XADFinderFlags;
 
 @class UniversalDetector;
 
-@interface XADArchive:NSObject
-{
-	XADArchiveParser *parser;
-	XADUnarchiver *unarchiver;
+@interface XADArchive : NSObject {
+    XADArchiveParser *parser;
+    XADUnarchiver *unarchiver;
 
-	id delegate;
-	NSTimeInterval update_interval;
-	XADError lasterror;
+    id delegate;
+    NSTimeInterval update_interval;
+    XADError lasterror;
 
-	NSMutableArray *dataentries,*resourceentries;
-	NSMutableDictionary *namedict;
+    NSMutableArray *dataentries, *resourceentries;
+    NSMutableDictionary *namedict;
 
-	off_t extractsize,totalsize;
-	int extractingentry;
-	BOOL extractingresource;
-	NSString *immediatedestination;
-	BOOL immediatesubarchives,immediatefailed;
-	off_t immediatesize;
-	XADArchive *parentarchive;
+    off_t extractsize, totalsize;
+    int extractingentry;
+    BOOL extractingresource;
+    NSString *immediatedestination;
+    BOOL immediatesubarchives, immediatefailed;
+    off_t immediatesize;
+    XADArchive *parentarchive;
 }
 
-+(XADArchive *)archiveForFile:(NSString *)filename;
-+(XADArchive *)recursiveArchiveForFile:(NSString *)filename;
++ (XADArchive *)archiveForFile:(NSString *)filename;
++ (XADArchive *)recursiveArchiveForFile:(NSString *)filename;
 
 
 
--(id)init;
--(id)initWithFile:(NSString *)file;
--(id)initWithFile:(NSString *)file error:(XADError *)error;
--(id)initWithFile:(NSString *)file delegate:(id)del error:(XADError *)error;
--(id)initWithData:(NSData *)data;
--(id)initWithData:(NSData *)data error:(XADError *)error;
--(id)initWithData:(NSData *)data delegate:(id)del error:(XADError *)error;
--(id)initWithArchive:(XADArchive *)archive entry:(int)n;
--(id)initWithArchive:(XADArchive *)archive entry:(int)n error:(XADError *)error;
--(id)initWithArchive:(XADArchive *)otherarchive entry:(int)n delegate:(id)del error:(XADError *)error;
--(id)initWithArchive:(XADArchive *)otherarchive entry:(int)n
-     immediateExtractionTo:(NSString *)destination error:(XADError *)error;
--(id)initWithArchive:(XADArchive *)otherarchive entry:(int)n
-     immediateExtractionTo:(NSString *)destination subArchives:(BOOL)sub error:(XADError *)error;
--(void)dealloc;
+- (id)init;
+- (id)initWithFile:(NSString *)file;
+- (id)initWithFile:(NSString *)file error:(XADError *)error;
+- (id)initWithFile:(NSString *)file delegate:(id)del error:(XADError *)error;
+- (id)initWithData:(NSData *)data;
+- (id)initWithData:(NSData *)data error:(XADError *)error;
+- (id)initWithData:(NSData *)data delegate:(id)del error:(XADError *)error;
+- (id)initWithArchive:(XADArchive *)archive entry:(int)n;
+- (id)initWithArchive:(XADArchive *)archive entry:(int)n error:(XADError *)error;
+- (id)initWithArchive:(XADArchive *)otherarchive entry:(int)n delegate:(id)del error:(XADError *)error;
+- (id)    initWithArchive:(XADArchive *)otherarchive entry:(int)n
+    immediateExtractionTo:(NSString *)destination error:(XADError *)error;
+- (id)    initWithArchive:(XADArchive *)otherarchive entry:(int)n
+    immediateExtractionTo:(NSString *)destination subArchives:(BOOL)sub error:(XADError *)error;
+- (void)dealloc;
 
--(BOOL)_parseWithErrorPointer:(XADError *)error;
+- (BOOL)_parseWithErrorPointer:(XADError *)error;
 
--(NSString *)filename;
--(NSArray *)allFilenames;
--(NSString *)formatName;
--(BOOL)isEncrypted;
--(BOOL)isSolid;
--(BOOL)isCorrupted;
--(int)numberOfEntries;
--(BOOL)immediateExtractionFailed;
--(NSString *)commonTopDirectory;
--(NSString *)comment;
+- (NSString *)filename;
+- (NSArray *) allFilenames;
+- (NSString *)formatName;
+- (BOOL)      isEncrypted;
+- (BOOL)      isSolid;
+- (BOOL)      isCorrupted;
+- (int)       numberOfEntries;
+- (BOOL)      immediateExtractionFailed;
+- (NSString *)commonTopDirectory;
+- (NSString *)comment;
 
--(void)setDelegate:(id)delegate;
--(id)delegate;
+- (void)setDelegate:(id)delegate;
+- (id)  delegate;
 
--(NSString *)password;
--(void)setPassword:(NSString *)newpassword;
+- (NSString *)password;
+- (void)      setPassword:(NSString *)newpassword;
 
--(NSStringEncoding)nameEncoding;
--(void)setNameEncoding:(NSStringEncoding)encoding;
+- (NSStringEncoding)nameEncoding;
+- (void)            setNameEncoding:(NSStringEncoding)encoding;
 
--(XADError)lastError;
--(void)clearLastError;
--(NSString *)describeLastError;
--(NSString *)describeError:(XADError)error;
+- (XADError)  lastError;
+- (void)      clearLastError;
+- (NSString *)describeLastError;
+- (NSString *)describeError:(XADError)error;
 
--(NSString *)description;
+- (NSString *)description;
 
 
 
--(NSDictionary *)dataForkParserDictionaryForEntry:(int)n;
--(NSDictionary *)resourceForkParserDictionaryForEntry:(int)n;
--(NSDictionary *)combinedParserDictionaryForEntry:(int)n;
+- (NSDictionary *)dataForkParserDictionaryForEntry:(NSInteger)n;
+- (NSDictionary *)resourceForkParserDictionaryForEntry:(NSInteger)n;
+- (NSDictionary *)combinedParserDictionaryForEntry:(NSInteger)n;
 
--(NSString *)nameOfEntry:(int)n;
--(XADPath *)rawNameOfEntry:(int)n;
--(BOOL)entryHasSize:(int)n;
--(off_t)uncompressedSizeOfEntry:(int)n;
--(off_t)compressedSizeOfEntry:(int)n;
--(off_t)representativeSizeOfEntry:(int)n;
--(BOOL)entryIsDirectory:(int)n;
--(BOOL)entryIsLink:(int)n;
--(BOOL)entryIsEncrypted:(int)n;
--(BOOL)entryIsArchive:(int)n;
--(BOOL)entryHasResourceFork:(int)n;
--(NSString *)commentForEntry:(int)n;
--(NSDictionary *)attributesOfEntry:(int)n;
--(NSDictionary *)attributesOfEntry:(int)n withResourceFork:(BOOL)resfork;
--(CSHandle *)handleForEntry:(int)n;
--(CSHandle *)handleForEntry:(int)n error:(XADError *)error;
--(CSHandle *)resourceHandleForEntry:(int)n;
--(CSHandle *)resourceHandleForEntry:(int)n error:(XADError *)error;
--(NSData *)contentsOfEntry:(int)n;
--(NSData *)contentsOfEntry:(int)n withLength:(NSInteger)length;
+- (NSString *)nameOfEntry:(NSInteger)n;
+- (XADPath *)rawNameOfEntry:(NSInteger)n;
+- (BOOL)entryHasSize:(NSInteger)n;
+- (off_t)uncompressedSizeOfEntry:(NSInteger)n;
+- (off_t)compressedSizeOfEntry:(NSInteger)n;
+- (off_t)representativeSizeOfEntry:(NSInteger)n;
+- (BOOL)entryIsDirectory:(NSInteger)n;
+- (BOOL)entryIsLink:(NSInteger)n;
+- (BOOL)entryIsEncrypted:(NSInteger)n;
+- (BOOL)entryIsArchive:(NSInteger)n;
+- (BOOL)entryHasResourceFork:(NSInteger)n;
+- (NSString *)commentForEntry:(NSInteger)n;
+- (NSDictionary *)attributesOfEntry:(NSInteger)n;
+- (NSDictionary *)attributesOfEntry:(NSInteger)n withResourceFork:(BOOL)resfork;
+- (CSHandle *)handleForEntry:(NSInteger)n;
+- (CSHandle *)handleForEntry:(NSInteger)n error:(XADError *)error;
+- (CSHandle *)resourceHandleForEntry:(NSInteger)n;
+- (CSHandle *)resourceHandleForEntry:(NSInteger)n error:(XADError *)error;
+- (NSData *)contentsOfEntry:(NSInteger)n;
+- (NSData *)contentsOfEntry:(NSInteger)n withLength:(NSInteger)length;
 //-(NSData *)resourceContentsOfEntry:(int)n;
 
--(BOOL)extractTo:(NSString *)destination;
--(BOOL)extractTo:(NSString *)destination subArchives:(BOOL)sub;
--(BOOL)extractEntries:(NSIndexSet *)entryset to:(NSString *)destination;
--(BOOL)extractEntries:(NSIndexSet *)entryset to:(NSString *)destination subArchives:(BOOL)sub;
--(BOOL)extractEntry:(int)n to:(NSString *)destination;
--(BOOL)extractEntry:(int)n to:(NSString *)destination deferDirectories:(BOOL)defer;
--(BOOL)extractEntry:(int)n to:(NSString *)destination deferDirectories:(BOOL)defer
-resourceFork:(BOOL)resfork;
--(BOOL)extractEntry:(int)n to:(NSString *)destination deferDirectories:(BOOL)defer
-dataFork:(BOOL)datafork resourceFork:(BOOL)resfork;
+- (BOOL)extractTo:(NSString *)destination;
+- (BOOL)extractTo:(NSString *)destination subArchives:(BOOL)sub;
+- (BOOL)extractEntries:(NSIndexSet *)entryset to:(NSString *)destination;
+- (BOOL)extractEntries:(NSIndexSet *)entryset to:(NSString *)destination subArchives:(BOOL)sub;
+- (BOOL)extractEntry:(int)n to:(NSString *)destination;
+- (BOOL)extractEntry:(int)n to:(NSString *)destination deferDirectories:(BOOL)defer;
+- (BOOL)extractEntry:(int)n to:(NSString *)destination deferDirectories:(BOOL)defer
+        resourceFork:(BOOL)resfork;
+- (BOOL)extractEntry:(int)n to:(NSString *)destination deferDirectories:(BOOL)defer
+            dataFork:(BOOL)datafork resourceFork:(BOOL)resfork;
 //-(BOOL)extractEntry:(int)n to:(NSString *)destination deferDirectories:(BOOL)defer
 //dataFork:(BOOL)datafork resourceFork:(BOOL)resfork;
--(BOOL)extractArchiveEntry:(int)n to:(NSString *)destination;
+- (BOOL)extractArchiveEntry:(int)n to:(NSString *)destination;
 
--(BOOL)_extractEntry:(int)n as:(NSString *)destfile deferDirectories:(BOOL)defer
-dataFork:(BOOL)datafork resourceFork:(BOOL)resfork;
+- (BOOL)_extractEntry:(int)n as:(NSString *)destfile deferDirectories:(BOOL)defer
+             dataFork:(BOOL)datafork resourceFork:(BOOL)resfork;
 
--(void)updateAttributesForDeferredDirectories;
+- (void)updateAttributesForDeferredDirectories;
 
 //Tim Oliver
 - (BOOL)extractContentsOfEntry:(int)n toPath:(NSString *)destination;
 
 // Deprecated
 
-+(NSArray *)volumesForFile:(NSString *)filename;
++ (NSArray *)volumesForFile:(NSString *)filename;
 
--(int)sizeOfEntry:(int)n;
--(void *)xadFileInfoForEntry:(int)n;
--(BOOL)extractEntry:(int)n to:(NSString *)destination overrideWritePermissions:(BOOL)override;
--(BOOL)extractEntry:(int)n to:(NSString *)destination overrideWritePermissions:(BOOL)override resourceFork:(BOOL)resfork;
--(void)fixWritePermissions;
+- (int)sizeOfEntry:(int)n;
+- (void *)xadFileInfoForEntry:(int)n;
+- (BOOL)extractEntry:(int)n to:(NSString *)destination overrideWritePermissions:(BOOL)override;
+- (BOOL)extractEntry:(int)n to:(NSString *)destination overrideWritePermissions:(BOOL)override resourceFork:(BOOL)resfork;
+- (void)  fixWritePermissions;
 
 @end
 
@@ -157,29 +156,29 @@ dataFork:(BOOL)datafork resourceFork:(BOOL)resfork;
 
 @interface NSObject (XADArchiveDelegate)
 
--(NSStringEncoding)archive:(XADArchive *)archive encodingForData:(NSData *)data guess:(NSStringEncoding)guess confidence:(float)confidence;
--(XADAction)archive:(XADArchive *)archive nameDecodingDidFailForEntry:(int)n data:(NSData *)data;
+- (NSStringEncoding)archive:(XADArchive *)archive encodingForData:(NSData *)data guess:(NSStringEncoding)guess confidence:(float)confidence;
+- (XADAction)archive:(XADArchive *)archive nameDecodingDidFailForEntry:(int)n data:(NSData *)data;
 
--(BOOL)archiveExtractionShouldStop:(XADArchive *)archive;
--(BOOL)archive:(XADArchive *)archive shouldCreateDirectory:(NSString *)directory;
--(XADAction)archive:(XADArchive *)archive entry:(int)n collidesWithFile:(NSString *)file newFilename:(NSString **)newname;
--(XADAction)archive:(XADArchive *)archive entry:(int)n collidesWithDirectory:(NSString *)file newFilename:(NSString **)newname;
--(XADAction)archive:(XADArchive *)archive creatingDirectoryDidFailForEntry:(int)n;
+- (BOOL)archiveExtractionShouldStop:(XADArchive *)archive;
+- (BOOL)archive:(XADArchive *)archive shouldCreateDirectory:(NSString *)directory;
+- (XADAction)archive:(XADArchive *)archive entry:(int)n collidesWithFile:(NSString *)file newFilename:(NSString * *)newname;
+- (XADAction)archive:(XADArchive *)archive entry:(int)n collidesWithDirectory:(NSString *)file newFilename:(NSString * *)newname;
+- (XADAction)archive:(XADArchive *)archive creatingDirectoryDidFailForEntry:(int)n;
 
--(void)archiveNeedsPassword:(XADArchive *)archive;
+- (void)archiveNeedsPassword:(XADArchive *)archive;
 
--(void)archive:(XADArchive *)archive extractionOfEntryWillStart:(int)n;
--(void)archive:(XADArchive *)archive extractionProgressForEntry:(int)n bytes:(off_t)bytes of:(off_t)total;
--(void)archive:(XADArchive *)archive extractionOfEntryDidSucceed:(int)n;
--(XADAction)archive:(XADArchive *)archive extractionOfEntryDidFail:(int)n error:(XADError)error;
--(XADAction)archive:(XADArchive *)archive extractionOfResourceForkForEntryDidFail:(int)n error:(XADError)error;
+- (void)archive:(XADArchive *)archive extractionOfEntryWillStart:(int)n;
+- (void)archive:(XADArchive *)archive extractionProgressForEntry:(int)n bytes:(off_t)bytes of:(off_t)total;
+- (void)archive:(XADArchive *)archive extractionOfEntryDidSucceed:(int)n;
+- (XADAction)archive:(XADArchive *)archive extractionOfEntryDidFail:(int)n error:(XADError)error;
+- (XADAction)archive:(XADArchive *)archive extractionOfResourceForkForEntryDidFail:(int)n error:(XADError)error;
 
--(void)archive:(XADArchive *)archive extractionProgressBytes:(off_t)bytes of:(off_t)total;
--(void)archive:(XADArchive *)archive extractionProgressFiles:(int)files of:(int)total;
+- (void)archive:(XADArchive *)archive extractionProgressBytes:(off_t)bytes of:(off_t)total;
+- (void)archive:(XADArchive *)archive extractionProgressFiles:(int)files of:(int)total;
 
 // Deprecated
--(NSStringEncoding)archive:(XADArchive *)archive encodingForName:(const char *)bytes guess:(NSStringEncoding)guess confidence:(float)confidence;
--(XADAction)archive:(XADArchive *)archive nameDecodingDidFailForEntry:(int)n bytes:(const char *)bytes;
+- (NSStringEncoding)archive:(XADArchive *)archive encodingForName:(const char *)bytes guess:(NSStringEncoding)guess confidence:(float)confidence;
+- (XADAction)archive:(XADArchive *)archive nameDecodingDidFailForEntry:(int)n bytes:(const char *)bytes;
 
 @end
